+++ frontend/src/components/Navigation/Navigation.css
@@ -1,0 +1,9 @@
+.nav-list {
+
+    background-color: rgb(0, 255, 30);
+}
+
+.nav-list li {
+    border: solid black 1px;
+    padding: 5px;
+}
