+++ frontend/src/components/Navigation/ProfileButton.js
@@ -1,11 +1,33 @@
-export const ProfileButton = () => {
-    return (<div style={
-        {
-            color: 'purple', fontSize: "65px"
-        }
-    }
+// frontend/src/components/Navigation/ProfileButton.js
+import React, { useState } from "react";
+import { useDispatch } from 'react-redux';
+import * as sessionActions from '../../store/session';
 
-    > <i className="fa-regular fa-user"></i> </div>);
+function ProfileButton({ user }) {
+    const dispatch = useDispatch();
+
+    const logout = (e) => {
+        e.preventDefault();
+        dispatch(sessionActions.logout());
+    };
+
+    const ulClassName = "profile-dropdown";
+
+    return (
+        <>
+            <button>
+                <i className="fas fa-user-circle" />
+            </button>
+            <ul className="profile-dropdown">
+                <li>{user.username}</li>
+                <li>{user.firstName} {user.lastName}</li>
+                <li>{user.email}</li>
+                <li>
+                    <button onClick={logout}>Log Out</button>
+                </li>
+            </ul>
+        </>
+    );
 }
 
-    ;
+export default ProfileButton;
