+++ frontend/src/components/Navigation/index.js
@@ -1,26 +1,18 @@
 // frontend/src/components/Navigation/index.js
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import { useSelector, useDispatch } from 'react-redux';
-import { ProfileButton } from './ProfileButton';
-import * as sessionActions from '../../store/session';
+import { useSelector } from 'react-redux';
+import ProfileButton from './ProfileButton';
 import './Navigation.css';
 
 function Navigation({ isLoaded }) {
     const sessionUser = useSelector(state => state.session.user);
-    const dispatch = useDispatch();
-
-    const logout = (e) => {
-        e.preventDefault();
-        dispatch(sessionActions.logout());
-    };
 
     let sessionLinks;
     if (sessionUser) {
         sessionLinks = (
             <li>
                 <ProfileButton user={sessionUser} />
-                <button onClick={logout}>Log Out</button>
             </li>
         );
     } else {
@@ -33,7 +25,7 @@ function Navigation({ isLoaded }) {
     }
 
     return (
-        <ul>
+        <ul className='nav-list'>
             <li>
                 <NavLink exact to="/">Home</NavLink>
             </li>
